:recycle: [refactor] selectBox switch문 적용
