--- src/components/common/Responsive.tsx
@@ -1,25 +1,0 @@
-import styled from "styled-components";
-
-interface ResponsiveProps {
-  [key: string]: any;
-}
-
-const ResponsiveWrapper = styled.div`
-  padding-left: 1rem;
-  padding-right: 1rem;
-  width: 1024px;
-  margin: 0 auto;
-
-  @media (max-width: 1024px) {
-    width: 768px;
-  }
-  @media (max-width: 768px) {
-    width: 100%;
-  }
-`;
-
-const Responsive = ({ children, ...rest }: ResponsiveProps) => {
-  return <ResponsiveWrapper {...rest}>{children}</ResponsiveWrapper>;
-};
-
-export default Responsive;
